Create PMC pub/sub mechanism
