+++ startup/titus-isolate
@@ -13,9 +13,11 @@ from titus_isolate.docker.event_manager import EventManager
 from titus_isolate.docker.free_event_handler import FreeEventHandler
 from titus_isolate.docker.utils import get_current_workloads
 from titus_isolate.isolate.workload_manager import WorkloadManager
+from titus_isolate.metrics.atlas_pmc_subscriber import AtlasPmcSubscriber
 from titus_isolate.metrics.bpf import get_bpf
 from titus_isolate.metrics.linux_context_provider import LinuxContextProvider
-from titus_isolate.metrics.metrics_reporter import MetricsReporter
+from titus_isolate.metrics.internal_metrics_reporter import InternalMetricsReporter
+from titus_isolate.metrics.metrics_manager import MetricsManager
 from titus_isolate.metrics.pmc_provider import PmcProvider
 from titus_isolate.metrics.pmc_publisher import PmcPublisher
 from titus_isolate.model.processor.config import get_cpu_from_env
@@ -47,13 +49,15 @@ def main(admin_port):
     event_manager = EventManager(docker.from_env().events(), event_handlers)
     set_em(event_manager)
 
-    # Report metrics
-    log.info("Starting metrics reporting...")
-    MetricsReporter(workload_manager, event_manager)
-
     # Publish PMC metrics
     log.info("Starting PMC publishing...")
-    PmcPublisher(PmcProvider(get_bpf(), LinuxContextProvider()))
+    pmc_subscriber = AtlasPmcSubscriber()
+    PmcPublisher(PmcProvider(get_bpf(), LinuxContextProvider()), [pmc_subscriber])
+
+    # Report metrics
+    log.info("Starting metrics reporting...")
+    internal_reporter = InternalMetricsReporter(workload_manager, event_manager)
+    MetricsManager([pmc_subscriber, internal_reporter])
 
     # Initialize currently running containers as workloads
     log.info("Isolating currently running workloads...")
