+++ titus_isolate/metrics/pmc_subscriber.py
@@ -1,0 +1,9 @@
+import collections
+
+
+class PmcSubscriber:
+    def __init__(self):
+        self.q = collections.deque()
+
+    def handle_nowait(self, metric):
+        self.q.appendleft(metric)
